+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+1.8.10
+
+* added contex menu "Explain" with current page context
+* changed @{{page}} to represent content
+* added @{{source}} to represent html
+
 1.7.30
 
 * added user command import and export